+++ src/renderer/components/ObservationCard.module.scss
@@ -36,10 +36,6 @@ $card-shadow-selected:
       min-height: 40px;
       padding: 5px;
 
-      .predictionOverride {
-        @include mix.custom-react-select;
-      }
-
       .selectButton {
         background: none;
         opacity: 0;
