+++ src/renderer/components/ObservationsHeader.tsx
@@ -2,12 +2,10 @@ import { Button } from '@blueprintjs/core';
 import classNames from 'classnames/bind';
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { StylesConfig } from 'react-select';
-import CreatableSelect from 'react-select/creatable';
 
 import { taxonOptions } from '../constants/taxons';
+import CreatableSelect from './CreatableSelect';
 import styles from './ObservationsHeader.module.scss';
-import sassVariables from '../styles/variables.module.scss';
 
 const cx = classNames.bind(styles);
 
@@ -56,61 +54,6 @@ export default function ObservationsHeader(props: ObservationsHeaderProps) {
     ? t('explore.inspect.selected', { count: selectedCardsTotal })
     : t('explore.inspect.header', { station: observations[0].station });
 
-  const customStyles: StylesConfig = {
-    menu: (base) => ({
-      ...base,
-      width: 'auto',
-    }),
-    option: (base, { data, isDisabled, isFocused, isSelected }) => ({
-      ...base,
-      // eslint-disable-next-line no-nested-ternary
-      backgroundColor: isSelected
-        ? sassVariables.green80
-        : isFocused
-        ? sassVariables.green10
-        : sassVariables.white,
-      borderBottom: `1px solid ${sassVariables.green10}`,
-      color: isSelected ? sassVariables.white : sassVariables.green80,
-    }),
-    control: (base) => ({
-      ...base,
-      background: sassVariables.green10,
-      border: 'none',
-      borderRadius: sassVariables.radiusInput,
-      boxShadow: 'none',
-      cursor: 'pointer',
-      display: 'flex',
-      fontWeight: 600,
-      minHeight: 0,
-      padding: '5px',
-      width: 200,
-    }),
-    valueContainer: (base) => ({
-      ...base,
-      padding: '0 5px 0 0',
-    }),
-    singleValue: (base) => ({
-      ...base,
-      color: sassVariables.green80,
-    }),
-    placeholder: () => ({
-      color: sassVariables.green60,
-    }),
-    input: (base) => ({
-      ...base,
-      margin: 0,
-      padding: 0,
-    }),
-    indicatorSeparator: (base) => ({
-      ...base,
-      display: 'none',
-    }),
-    indicatorsContainer: (base) => ({
-      ...base,
-      padding: 0,
-    }),
-  };
-
   return (
     <div className={containerClass}>
       <div className={cx({ side: true, left: true })}>
@@ -131,8 +74,6 @@ export default function ObservationsHeader(props: ObservationsHeaderProps) {
               value={globalOverride}
               options={taxonOptions}
               onChange={setGlobalOverride}
-              className={styles.predictionOverride}
-              styles={customStyles}
               isDisabled={!selectedCardsTotal}
               isClearable={!!globalOverride}
               menuShouldScrollIntoView={false}
